+++ libraries/test_session_common/src/main/java/androidx/media3/test/session/common/MediaBrowserServiceCompatConstants.java
@@ -32,6 +32,8 @@ public class MediaBrowserServiceCompatConstants {
   public static final String TEST_SEND_CUSTOM_COMMAND = "sendCustomCommand";
   public static final String TEST_MEDIA_ITEMS_WITH_BROWSE_ACTIONS =
       "getLibraryRoot_withBrowseActions";
+  public static final String TEST_SUBSCRIBE_THEN_REJECT_ON_LOAD_CHILDREN =
+      "subscribe_thenRejectOnLoadChildren";
 
   private MediaBrowserServiceCompatConstants() {}
 }